// src/_bots/utils/types.ts
export type Author = 'user' | 'bot'
export type ToneType = 'Creative' | 'Balanced' | 'Precise'


export type SendMessageOptions = {
  conversationId?: string
  clientId?: string
  conversationSignature?: string
  conversationExpiryTime?: string
  invocationId?: string
  messageType?: string
  variant?: ToneType
  locale?: string
  market?: string
  region?: string
  location?: {
    lat: number | string
    lng: number | string
    re?: string
  }
  locationHints?: LocationHintChatRequestMessage

  onMessage?: (partialResponse: ChatMessage) => void
}

export interface ChatMessage {
  id: string
  text: string
  prompt?: string
  messageId?: string
  author: Author
  contentType?: 'IMAGE'
  conversationId: string
  clientId: string
  conversationSignature: string
  conversationExpiryTime?: string
  invocationId?: string
  messageType?: string
  variant?: ToneType

  detail?: ChatMessageFull | ChatMessagePartial
  end?: boolean
}

export interface ConversationResult {
  conversationId: string
  clientId: string
  conversationSignature: string
  result: APIResult
}

export interface APIResult {
  value: string
  message: null
}

export interface ChatUpdate {
  type: 1
  target: string
  arguments: ChatUpdateArgument[]
}

export interface ChatUpdateArgument {
  messages: ChatMessagePartial[]
  requestId: string
  result: null
}

export interface ChatMessagePartial {
  text: string
  author: Author
  createdAt: string
  timestamp: string
  messageId: string
  offense: string
  adaptiveCards: AdaptiveCard[]
  sourceAttributions: any[]
  feedback: ChatMessageFeedback
  contentOrigin: string
  privacy?: null
  messageType?: string
  suggestedResponses?: ChatMessagePartial[]
}

export interface AdaptiveCard {
  type: string
  version: string
  body: AdaptiveCardBody[]
}

export interface AdaptiveCardBody {
  type: string
  text: string
  wrap: boolean
}

export interface ChatMessageFeedback {
  tag: null
  updatedOn: null
  type: string
}

export interface ChatUpdateCompleteResponse {
  type: 2
  invocationId: string
  item: ChatResponseItem
}

export interface ChatResponseItem {
  messages: ChatMessageFull[]
  firstNewMessageIndex: number
  suggestedResponses: null
  conversationId: string
  requestId: string
  conversationExpiryTime: string
  telemetry: Telemetry
  result: ChatRequestResult
  throttling: Throttling
}

export interface ChatMessageFull {
  text: string
  contentType?: 'IMAGE'
  spokenText?: string
  author: Author
  from?: ChatMessageFrom
  createdAt: string
  timestamp: string
  locale?: string
  market?: string
  region?: string
  location?: string
  locationHints?: LocationHint[]
  messageId: string
  requestId: string
  offense: string
  feedback: ChatMessageFeedback
  contentOrigin: string
  privacy?: null
  inputMethod?: string
  adaptiveCards?: AdaptiveCard[]
  sourceAttributions?: any[]
  suggestedResponses?: SuggestedResponse[]
  messageType?: string
}

export interface Throttling {
  maxNumUserMessagesInConversation: number
  numUserMessagesInConversation: number
}

export interface ChatMessageFrom {
  id: string
  name: null
}

export interface LocationHint {
  country: string
  countryConfidence: number
  state: string
  city: string
  cityConfidence: number
  zipCode: string
  timeZoneOffset: number
  dma: number
  sourceType: number
  center: Coords
  regionType: number
}

export interface Coords {
  latitude: number
  longitude: number
  height: null
}

export interface SuggestedResponse {
  text: string
  messageId: string
  messageType: string
  contentOrigin: string
  author?: Author
  createdAt?: string
  timestamp?: string
  offense?: string
  feedback?: ChatMessageFeedback
  privacy?: null
}

export interface ChatRequestResult {
  value: string
  serviceVersion: string
}

export interface Telemetry {
  metrics?: null
  startTime: string
}
// ...
export interface LocationHintChatRequestMessage {
  country: string
  state: string
  city: string
  zipcode: string
  timezoneoffset: number
  dma: number
  countryConfidence: number
  cityConfidence: number
  Center: Center
  RegionType: number
  SourceType: number
}

export interface Center {
  Latitude: number
  Longitude: number
}